Trx file output for Chutzpah.console.exe
I have added rudimentary support for TRX file generation through
chutzpah.console.exe.
• VSTS Extensions: This file currently enum of “TestRunItemType”,  which
is the order of items stored in “TestRunType” which is serialized into
trx file. It also contains a method “GetInstace<T>( TestRunItemType
type) that returns the desired type from an object array.
• TrxConsoleRunnerCallback.cs: I have extended the exisisting console
runner to listen on events and log the same. TO-NOTE: I am still calling
the ConsoleRunner callback on all the methods, so the trace still works.
• VSTSTestCase.cs: that extends the exisisting TestCase. I needed couple
of more information about the runs like the start-time, end-time,
exception, Id, executionId etc.
• CommandLine.cs: I have added a parameter that specifies trx output
file location.
• Program.cs: I initialize the callback based on the command line
parameters.
• Vsts.xsd and vsts.cs: xsd is part of visual studio, vsts.cs is
generated. Although I couldn’t use the vsts.cs that was generated. There
were some bugs in that. I corrected the same. This is the schema that is
used by VS to view trx files.