I have added rudimentary support for TRX report
• VSTS Extensions: This file currently enum of “TestRunItemType”,  which
is the order of items stored in “TestRunType” which is serialized into
trx file. It also contains a method “GetInstace<T>( TestRunItemType
type) that returns the desired type from an object array.
• TrxXmlTransformer.cs: I have extended the exisisting transformer and
overrode the Transform function that returns Unicode string. (There are
some weird stuff, we can discuss this)
• VSTSTestCase.cs: that extends the exisisting TestCase. I needed couple
of more information about the runs like the start-time, end-time,
exception, Id, executionId etc.
•FileSystemWrapper.cs & SummaryTransformer.cs: I have added WriteAllText
with encode parameter, as visual studio trx xsd expects a Unicode char
base for xml whilst the exisiting junitXml needs utf-8, so i have added
encoding property(Field), so that it can be controlled by implementing
classes. viz. TrxXmlTransformer. I have defaulted the encoding to
UTF-8(I personally feel it should be unicode) but we can discuss on
that.
• SummaryTransformerFactory.cs: This was to just enable trx output
• Vsts.xsd and vsts.cs: xsd is part of visual studio, vsts.cs is
generated. Although I couldn’t use the vsts.cs that was generated. There
were some bugs in that. I corrected the same. This is the schema that is
used by VS to view trx files.