Update test.js
the outerHtml was needed in my code examples without your implementation of an script tag. But with the script tag and the ID for the script tag, we only need to get the inner HTML because with: 
    "document.getElementById("template_comment")" 
we get the Script Tag you create and for our template we only need the template code exclude the script tags.