Move 'browser' variable to outer scope
The browser variable is declared in the inner scope of a try-catch block.

There are references to this variable outside that block which may lead to a 'variable not defined' error (line 343/344 for example).

This leads to an issue when the execution of a test file exceeds the configured timeout. In this case the chutpah.console.exe will terminate with a generic 'exit code 2' from the general error handler in the qunitRunner.js component instead of the intended timeout error.